splitted ether.c in 3 (_common, _sunos, _nethub)
