+++ src/xc.c
